removed environment flag
