+++ src/applications/gi/containers/search/LocationSearchResults.jsx
@@ -4,23 +4,24 @@ import mapboxgl from 'mapbox-gl';
 import { focusElement, getScrollOptions } from 'platform/utilities/ui';
 import LoadingIndicator from '@department-of-veterans-affairs/component-library/LoadingIndicator';
 
-import ResultCard from '../../containers/search/ResultCard';
+import { connect } from 'react-redux';
+import classNames from 'classnames';
+import scrollTo from 'platform/utilities/ui/scrollTo';
+import recordEvent from 'platform/monitoring/record-event';
+import environment from 'platform/utilities/environment';
+import ResultCard from './ResultCard';
 import { mapboxToken } from '../../utils/mapboxToken';
 import { MapboxInit, MAX_SEARCH_AREA_DISTANCE, TABS } from '../../constants';
-import TuitionAndHousingEstimates from '../../containers/TuitionAndHousingEstimates';
-import FilterYourResults from '../../containers/FilterYourResults';
+import TuitionAndHousingEstimates from '../TuitionAndHousingEstimates';
+import FilterYourResults from '../FilterYourResults';
 import { createId } from '../../utils/helpers';
 import {
   fetchSearchByLocationCoords,
   updateEligibilityAndFilters,
   mapChanged,
 } from '../../actions';
-import { connect } from 'react-redux';
 import { getFiltersChanged } from '../../selectors/filters';
 import MobileFilterControls from '../../components/MobileFilterControls';
-import classNames from 'classnames';
-import scrollTo from 'platform/utilities/ui/scrollTo';
-import recordEvent from 'platform/monitoring/record-event';
 
 const MILE_METER_CONVERSION_RATE = 1609.34;
 const LIST_TAB = 'List';
@@ -40,7 +41,7 @@ function LocationSearchResults({
   const { count, results } = search.location;
   const { location, streetAddress } = search.query;
   const map = useRef(null);
-  const mapContainer = useRef(null);
+  // const mapContainer = useRef(null);
   const [markers, setMarkers] = useState([]);
   const [mapState, setMapState] = useState({ changed: false, distance: null });
   const [usedFilters, setUsedFilters] = useState(filtersChanged);
@@ -48,6 +49,20 @@ function LocationSearchResults({
   const [mobileTab, setMobileTab] = useState(LIST_TAB);
   const [markerClicked, setMarkerClicked] = useState(null);
   const [activeMarker, setActiveMarker] = useState(null);
+  const [myLocation, setMyLocation] = useState(null);
+  const usingUserLocation = () => {
+    if (environment.isProduction()) return true;
+
+    const currentPositions = document.getElementsByClassName(
+      'current-position',
+    );
+
+    if (currentPositions.length === 0) return false;
+    if (myLocation === null) setMyLocation(search.query.location);
+    if (search.query.location !== myLocation) return false;
+
+    return true;
+  };
 
   /**
    * When map is moved update distance from center to NorthEast corner
@@ -79,11 +94,13 @@ function LocationSearchResults({
    */
   const setupMap = () => {
     if (map.current) return; // initialize map only once
+    const container = document.getElementById('mapbox-gl-container');
+    if (!container) return;
 
     mapboxgl.accessToken = mapboxToken;
 
     const mapInit = new mapboxgl.Map({
-      container: mapContainer.current,
+      container: 'mapbox-gl-container',
       style: 'mapbox://styles/mapbox/outdoors-v11',
       center: [MapboxInit.centerInit.longitude, MapboxInit.centerInit.latitude],
       zoom: MapboxInit.zoomInit,
@@ -134,9 +151,9 @@ function LocationSearchResults({
    */
   useEffect(
     () => {
-      if (mapContainer.current) {
-        setupMap();
-      }
+      // if (mapContainer.current) {
+      setupMap();
+      // }
     },
     [mobileTab],
   );
@@ -266,6 +283,9 @@ function LocationSearchResults({
    */
   useEffect(
     () => {
+      map.current = null;
+      setupMap();
+
       markers.forEach(marker => marker.remove());
       setActiveMarker(null);
 
@@ -336,9 +356,11 @@ function LocationSearchResults({
         <span className="location-letter vads-u-font-size--sm">
           {index + 1}
         </span>
-        <span className="vads-u-padding-x--0p5 vads-u-font-size--sm">
-          <strong>{miles} miles</strong>
-        </span>
+        {usingUserLocation() && (
+          <span className="vads-u-padding-x--0p5 vads-u-font-size--sm">
+            <strong>{miles} miles</strong>
+          </span>
+        )}
       </div>
     );
 
@@ -435,7 +457,7 @@ function LocationSearchResults({
             </>
           )}
           {smallScreen && (
-            <MobileFilterControls className={'vads-u-margin-top--2'} />
+            <MobileFilterControls className="vads-u-margin-top--2" />
           )}
         </>
       );
@@ -586,7 +608,7 @@ function LocationSearchResults({
     return (
       <div className={containerClassNames}>
         <map
-          ref={mapContainer}
+          // ref={mapContainer}
           id="mapbox-gl-container"
           aria-label="Find VA locations on an interactive map"
           aria-describedby="map-instructions"
@@ -597,11 +619,11 @@ function LocationSearchResults({
             !smallScreen && (
               <div
                 id="search-area-control-container"
-                className={'mapboxgl-ctrl-top-center'}
+                className="mapboxgl-ctrl-top-center"
               >
                 <button
                   id="search-area-control"
-                  className={'usa-button'}
+                  className="usa-button"
                   onClick={searchArea}
                   disabled={!areaSearchWithinBounds}
                 >
@@ -622,7 +644,7 @@ function LocationSearchResults({
   // returns content ordered and setup for smallScreens
   if (smallScreen) {
     return (
-      <div className={'location-search vads-u-padding--1'}>
+      <div className="location-search vads-u-padding--1">
         {inProgress && <LoadingIndicator message="Loading search results..." />}
         {!inProgress && (
           <>
@@ -656,8 +678,8 @@ function LocationSearchResults({
 
   // Returns content setup for desktop screens
   return (
-    <div className={'location-search vads-u-padding-top--1'}>
-      <div className={'usa-width-one-third'}>
+    <div className="location-search vads-u-padding-top--1">
+      <div className="usa-width-one-third">
         {inProgress && <LoadingIndicator message="Loading search results..." />}
         {!inProgress && (
           <>
@@ -679,7 +701,7 @@ function LocationSearchResults({
         )}
       </div>
 
-      <div className={'usa-width-two-thirds'}>{mapElement()}</div>
+      <div className="usa-width-two-thirds">{mapElement()}</div>
     </div>
   );
 }
